parseIDToken moved to auth.js, authorizeUser partially moved to auth.js
