+++ src/components/Layout.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import Sidebar from './Sidebar';
 
 export default function Layout({
-  authorizeUser, children, filterFilesInAllAccounts, parseIDToken, removeAllAccounts, 
+  authorizeUser, children, filterFilesInAllAccounts, removeAllAccounts,
   starFilter, userList,
 }) {
   return (
@@ -11,8 +11,7 @@ export default function Layout({
       <Sidebar
         authorizeUser={authorizeUser}
         filterFilesInAllAccounts={filterFilesInAllAccounts}
-        parseIDToken={parseIDToken}
-        removeAllAccounts = {removeAllAccounts}
+        removeAllAccounts={removeAllAccounts}
         starFilter={starFilter}
         userList={userList}
       />
@@ -25,7 +24,6 @@ Layout.propTypes = {
   authorizeUser: PropTypes.func.isRequired,
   children: PropTypes.node.isRequired,
   filterFilesInAllAccounts: PropTypes.func.isRequired,
-  parseIDToken: PropTypes.func.isRequired,
   userList: PropTypes.arrayOf(PropTypes.object).isRequired,
-  removeAllAccountsAccounts: PropTypes.func.isRequired,
+  removeAllAccounts: PropTypes.func.isRequired,
 };
